+++ ngraph/nsphere.cpp
@@ -162,7 +162,7 @@ SpatialVertex** NSphere::findMNearestNeighbors(SpatialVertex* start){
 
 	for(int i = 0; i < m; i++){
 
-		dist[i] = DBL_MAX;
+		dist[i] = FLT_MAX;
 	}
 
 	for(int i = 0; i < N; i++){
@@ -255,7 +255,7 @@ void NSphere::equalize(){
 void NSphere::gradientDescent(float gamma, float tolerance, long int maxItr){
 
 	float* netForce[N];	// a record of the net force on each node
-	float prevMaxDisp = DBL_MAX;	// record of the maximum scalar displacement of a node on the previous iteration
+	float prevMaxDisp = FLT_MAX;	// record of the maximum scalar displacement of a node on the previous iteration
 									// initially set to an impossibly large value to ensure at least one iteration occurs
 	float tolMaxDisp = (radius * pow(N, -1.0/DIM) * tolerance) * baseGam;	// the maximum tolerated displacement
 																				// if the maximum displacement is above this value, continue iterating
@@ -421,15 +421,15 @@ NSphere* NSphere::importObject(const char* filename){
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
-	sscanf(line, "# radius = %lf", &radius);
+	sscanf(line, "# radius = %f", &radius);
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
-	sscanf(line, "# base gamma = %lf", &gamma);
+	sscanf(line, "# base gamma = %f", &gamma);
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
-	sscanf(line, "# base tolerance = %lf", &tolerance);
+	sscanf(line, "# base tolerance = %f", &tolerance);
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
@@ -445,7 +445,7 @@ NSphere* NSphere::importObject(const char* filename){
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
-	sscanf(line, "# iteration weights = %lf", &weights);
+	sscanf(line, "# iteration weights = %f", &weights);
 	memset(line, 0, bufsize * sizeof(char));
 	
 	infile.getline(line, bufsize);
